\name{lisrelToR-package}
\alias{lisrelToR-package}
\alias{lisrelToR}
\docType{package}
\title{
Import LISREL output in R.
}
\description{
This is an unofficial package aimed at automating the import of LISREL output in R. This package or its maintainer is not in any way affiliated with the creators of LISREL and SSI, Inc.
}

\author{
Sacha Epskamp (mail@sachaepskamp.com)

Maintainer: Sacha Epskamp <mail@sachaepskamp.com>
}
\references{
github.com/SachaEpskamp/lisrelToR
}

\keyword{ package }