\name{print.lisrel}
\alias{print.lisrel}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print method for lisrel object.
}
\description{
Prints LISREL output file to the console.
}
\usage{
\method{print}{lisrel}(x,...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
output of \code{\link{readLisrel}}
}
  \item{\dots}{Not used}
}
\author{
Sacha Epskamp <mail@sachaepskamp.com>
}
